+++ test/logger_test.js
@@ -53,4 +53,27 @@ describe('logger', function() {
 
     expect(logger.level()).to.equal(50);
   });
+
+  it('should not accept non-bunyan loggers without necessary logging functions', function() {
+    var nonBunyanLogger = {};
+
+    log.init({ logger: nonBunyanLogger });
+
+    var logger = log.getLogger('myLogger');
+    expect(logger).to.be.an.instanceOf(bunyan);
+  });
+
+  it('should accept non-bunyan loggers with necessary logging functions', function() {
+    var nonBunyanLogger = {
+      debug: function () {},
+      info: function () {},
+      warn: function () {},
+      error: function () {}
+    };
+
+    log.init({ logger: nonBunyanLogger });
+
+    var logger = log.getLogger('myLogger');
+    expect(logger).not.to.be.an.instanceOf(bunyan);
+  });
 });